Updated read from stdout logic, updated tests. Updated embedded openssh server start to wait for port to be opened

+++ embedded_server/openssh.py
@@ -1,6 +1,8 @@
 import os
+import socket
 
 from gevent.subprocess import Popen
+from gevent import sleep
 
 
 SERVER_KEY = os.path.sep.join([os.path.dirname(__file__), 'rsa.key'])
@@ -17,11 +19,17 @@ class OpenSSHServer(object):
                '-h', SERVER_KEY, '-f', SSHD_CONFIG]
         server = Popen(cmd)
         self.server_proc = server
+        self._wait_for_port()
+
+    def _wait_for_port(self):
+        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        while sock.connect_ex(('127.0.0.1', self.port)) != 0:
+            sleep(.1)
 
     def stop(self):
-        if self.server_proc is not None:
+        if self.server_proc is not None and self.server_proc.returncode is None:
             self.server_proc.terminate()
-            self.server_proc.communicate()
+            self.server_proc.wait()
 
     def __del__(self):
         self.stop()
